+++ NEWS.md
@@ -1,5 +1,7 @@
 # learnr (development version)
 
+-   Fixed an issue that prevented authors from using symbols, such as `T` or a variable, as the value of the `exercise` chunk option, which caused tutorials with chunks with `exercise = T` to fail to render (thanks @cknotz #757, #758).
+
 -   The embedded Ace editor used in learnr exercises now defaults to a tab width of 2, aligning with the Tidyverse style guide (#761).
 
 # learnr 0.11.2